sleep between ami deregister and snapshot delete
Sometimes the call to delete an image snapshot fails because the AMI
has not finished deregistering. Added some sleep time to decrease the
chances of this.
Really, this should be solved with a waiter in botocore.
Also moved the Backer initialization into the handler so we don't retain state
in the ec2 client between invocations.